backend/src/config.py
import os
from pathlib import Path
from dotenv import load_dotenv

# Load .env file if it exists (local dev only — on HF Spaces, use Secrets)
env_path = Path(__file__).resolve().parent.parent / ".env"
if env_path.exists():
    load_dotenv(dotenv_path=env_path)

# Database settings
DATABASE_URL = os.getenv("DATABASE_URL", "")
if not DATABASE_URL:
    raise ValueError("DATABASE_URL is not set. Set it in .env file or as an environment variable (HF Spaces Secrets).")

# JWT settings
SECRET_KEY = os.getenv("SECRET_KEY", "")
if not SECRET_KEY:
    raise ValueError("SECRET_KEY is not set. Set it in .env file or as an environment variable (HF Spaces Secrets).")

ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "30"))

# OpenAI API settings
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")


class Settings:
    """Application settings loaded from environment variables."""

    def __init__(self):
        self.database_url = DATABASE_URL
        self.secret_key = SECRET_KEY
        self.algorithm = ALGORITHM
        self.access_token_expire_minutes = ACCESS_TOKEN_EXPIRE_MINUTES
        self.openai_api_key = OPENAI_API_KEY


settings = Settings()
