backend/requirements.txt
# # fastapi==0.115.0
# # uvicorn==0.32.0
# # sqlmodel==0.0.22
# # pydantic==2.10.1
# # pydantic-settings==2.6.1
# # python-dotenv==1.0.1
# # psycopg2-binary==2.9.10
# # passlib==1.7.4
# # bcrypt==4.0.1
# # python-jose[cryptography]==3.3.0
# # python-multipart==0.0.20
# # better-exceptions>=0.3.0
# # alembic==1.13.3
# # asyncpg==0.30.0
# # mcp>=1.0.0
# # openai>=1.0.0


# fastapi==0.115.0
# uvicorn==0.32.0
# sqlmodel==0.0.22
# pydantic==2.10.1
# pydantic-settings==2.6.1
# python-dotenv==1.0.1

# # Database
# psycopg2-binary==2.9.10
# asyncpg==0.30.0
# alembic==1.13.3

# # Auth & Security
# passlib[bcrypt]==1.7.4
# bcrypt==4.0.1
# python-jose[cryptography]==3.3.0
# python-multipart==0.0.20

# # Helpful Tools
# better-exceptions==0.3.3
# mcp>=1.0.0
# openai>=1.0.0


fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
pydantic==2.10.1
pydantic-settings==2.6.1
python-dotenv==1.0.1

psycopg2-binary==2.9.10
asyncpg==0.30.0
alembic==1.13.3

passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20

better-exceptions==0.3.3
mcp>=1.0.0
openai>=1.0.0